(min):.1s animation, fix item numberin
