+++ projects/@ngx-dummy/accordion-simple/src/helpers/IItemToggler.ts
@@ -1,0 +1,4 @@
+export interface IToggleer {
+  itemId: number;
+  isOpen: boolean;
+}
