+++ projects/@ngx-dummy/accordion-simple/src/helpers/iconsbase64.ts
