+++ projects/@ngx-dummy/accordion-simple/src/helpers/index.ts
@@ -1,0 +1,7 @@
+export {
+  IAccordionItemBodyStyling,
+  IAccordionItemHeadStyling,
+  IAccordionItemSpacingStyling,
+  IAccordionItemStyling,
+  IAccordionStyling
+} from './IAccordionStylings';
