+++ projects/@ngx-dummy/accordion-simple/src/lib/IAccordion.ts
@@ -5,9 +5,9 @@ export interface Accordion {
   openedItem?: number | string | null;
 }
 
-export interface AccordionItem {
+export type AccordionItem = {
   id: number | null;
   title: string;
   body: string;
   isOpen?: boolean;
-}
+} & { itemNum?: number | string | null }
