+++ projects/@ngx-dummy/accordion-simple/src/lib/accordion.component.html
@@ -1,3 +1,3 @@
 <section class="accordion">
-  <accord-simple-accordion-item *ngFor="let item of accordionList.items; let i = index" [item]="item" [styling]="itemStyle" (toggler)="onItemToggled($event)"></accord-simple-accordion-item>
+  <accord-simple-accordion-item *ngFor="let item of accordionList.items; let i = index" [item]="item" [styling]="itemStyle" (toggled)="onItemToggled($event)"></accord-simple-accordion-item>
 </section>
