+++ projects/@ngx-dummy/accordion-simple/src/lib/accordion.component.ts
@@ -1,8 +1,9 @@
 import { Component, OnInit, Input, HostBinding, Renderer2, ElementRef } from '@angular/core';
 
-import { dummyAccordionList } from './helpers/dummy-data';
-import { IAccordionItemStyling, IAccordionStyling } from './helpers/IAccordionStylings';
+import { dummyAccordionList } from '../helpers/dummy-data';
+import { IAccordionItemStyling, IAccordionStyling } from '../helpers/IAccordionStylings';
 import { Accordion } from './IAccordion';
+import { IToggleer } from "../helpers/IItemToggler";
 
 @Component({
   selector: 'accord-simple-accordion',
@@ -17,14 +18,12 @@ export class AccordionComponent implements OnInit {
     maxWidth: '200px',
     guttedItems: '1rem',
     itemStyling: {
-      headBgColor: 'teal',
+      headBgColor: '#4197b2',
       headColor: '#fff',
       bodyBgColor: '#fff',
       bodyColor: '#000',
       margin: '0',
-      padding: '0',
-      logo: '../assets/list-box.svg',
-      openSign: '../assets/plus.svg'
+      padding: '0'
     }
   };
   itemStyle: IAccordionItemStyling = {
@@ -36,19 +35,20 @@ export class AccordionComponent implements OnInit {
 
   ngOnInit() {
     const el = this.el.nativeElement;
-    const itemsGutts = (typeof this.accordionStyling.guttedItems === "boolean") ? null : '1rem';
+    const itemsGutts = (typeof this.accordionStyling.guttedItems == 'boolean' && this.accordionStyling.guttedItems === false) ? null : this.accordionStyling.guttedItems || '.5rem';
     this.itemStyle = {
       padding: '1px',
       margin: (itemsGutts && typeof itemsGutts === 'string') ? itemsGutts : '0',
       ...this.itemStyle,
       ...this.accordionStyling.itemStyling
     };
-    this.id = `accordion_${this.accordionList.id.toString()}`;
+    this.id = `accordion_${this.accordionList?.id?.toString() ?? '0'}`;
     this.accordionStyling.maxWidth && this.render.setStyle(el, 'max-width', this.accordionStyling.maxWidth);
   }
 
-  onItemToggled({ id, isOpen }: { id: number, isOpen: boolean; } = { id: 0, isOpen: false }) {
-    this.accordionList.items = this.accordionList?.items?.map(item => { return (item.id == id) ? { ...item, isOpen } : { ...item, isOpen: false }; });
+  onItemToggled({ itemId, isOpen }: IToggleer = { itemId: 0, isOpen: false }) {
+    console.log(`clicked item ${itemId}`)
+    this.accordionList.items = this.accordionList?.items?.map(item => { return (item.id == itemId) ? { ...item, isOpen } : { ...item, isOpen: false }; });
   }
 
 }
