+++ projects/@ngx-dummy/accordion-simple/src/lib/index.ts
@@ -2,10 +2,3 @@ export * from './accordion.component';
 export * from './accordion-item.component';
 export { Accordion, AccordionItem } from './IAccordion';
 export { AccordionModule } from './accordion.module';
-export {
-  IAccordionItemBodyStyling,
-  IAccordionItemHeadStyling,
-  IAccordionItemSpacingStyling,
-  IAccordionItemStyling,
-  IAccordionStyling
-} from './helpers/IAccordionStylings'
