+++ projects/@ngx-dummy/accordion-simple/src/public-api.ts
@@ -2,4 +2,7 @@
  * Public API Surface of accordion-simple-lib
  */
 
-export * from './lib/'
+export * from './lib/accordion.component';
+export * from './lib/accordion-item.component';
+export { Accordion, AccordionItem } from './lib/IAccordion';
+export { AccordionModule } from './lib/accordion.module';
